Rename pkg/fetcher to pkg/upstream and split code into more files
